+++ client/src/components/SearchedBookResults/SearchedBookResults.jsx
@@ -2,8 +2,33 @@ import { Container, Col, Card, Row, Button } from "react-bootstrap";
 import "./SearchedBookResults.scss";
 import auth from "../../utils/auth";
 import ModalBookDescription from "../Modal-BookDescription/ModalBookDescription";
+import { FAV_BOOK } from "../../utils/mutations";
+import { useMutation } from "@apollo/client";
 
-function SearchedBookResults({ searchedBooks, favBookIds, handleFavBook }) {
+function SearchedBookResults({ searchedBooks, favBookIds }) {
+
+  const [addBook, { error }] = useMutation(FAV_BOOK, {
+    refetchQueries: ["me"],
+  });
+
+  const handleFavBook = async (googleBookId) => {
+    console.log("Book Info: ", googleBookId);
+    const bookToFavorite = searchedBooks.find(
+      (book) => book.googleBookId === googleBookId
+    );
+    console.log("booktofavorite: ", bookToFavorite);
+
+    // console.log("userdata: ", userData);
+
+    try {
+      const response = await addBook({
+        variables: { book: bookToFavorite },
+      });
+      console.log("response from addBook: ", response);
+    } catch (e) {
+      console.log(e);
+    }
+  };
   return (
     <Container>
       <h2 className="pt-5">
@@ -14,7 +39,7 @@ function SearchedBookResults({ searchedBooks, favBookIds, handleFavBook }) {
       <Row>
         {searchedBooks.map((book) => {
           return (
-            <Col md="4" key={book.bookId}>
+            <Col md="4" key={book.googleBookId}>
               <Card className="custom-card" border="dark">
                 <Card.Title>{book.title}</Card.Title>
                 <ModalBookDescription
@@ -42,12 +67,12 @@ function SearchedBookResults({ searchedBooks, favBookIds, handleFavBook }) {
                   {auth.loggedIn() && (
                     <Button
                       disabled={favBookIds?.some(
-                        (favoritedBookId) => favoritedBookId === book.bookId
+                        (favoritedBookId) => favoritedBookId === book.googleBookId
                       )}
                       className="btn-block btn-info"
                       onClick={() => handleFavBook(book.bookId)}>
                       {favBookIds?.some(
-                        (favoritedBookId) => favoritedBookId === book.bookId
+                        (favoritedBookId) => favoritedBookId === book.googleBookId
                       )
                         ? "Favorited"
                         : "Add to Favorites"}
